+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Added
+
+- `client_id` now will be returned as a header, by default in `X-Ext-Auth-Client-Id` header.
+- Add `--client-id-header` CMD flag to customize the header that will be returned with the authenticated client.
+
 ## [v0.3.0] - 2022-08-01
 
 ### Added