+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [v0.4.0] - 2022-12-10
+
 ### Added
 
 - `client_id` now will be returned as a header, by default in `X-Ext-Auth-Client-Id` header.
@@ -33,7 +35,8 @@
 - Add token review Prometheus metrics.
 - Make configuration API public as a go library.
 
-[unreleased]: https://github.com/slok/simple-ingress-external-auth/compare/v0.3.0...HEAD
+[unreleased]: https://github.com/slok/simple-ingress-external-auth/compare/v0.4.0...HEAD
+[v0.4.0]: https://github.com/slok/simple-ingress-external-auth/compare/v0.3.0...v0.4.0
 [v0.3.0]: https://github.com/slok/simple-ingress-external-auth/compare/v0.2.0...v0.3.0
 [v0.2.0]: https://github.com/slok/simple-ingress-external-auth/compare/v0.1.0...v0.2.0
 [v0.1.0]: https://github.com/slok/simple-ingress-external-auth/releases/tag/v0.1.0